requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.9
importlib-metadata==1.6.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
importlib-metadata==1.6.1
pyTelegramBotAPI==3.7.1
python-dotenv==0.13.0
requests==2.7.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
zipp==3.1.0